+++ test/recording_studio_wordpress_plugin_template_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class RecordingStudioWordpressPluginTemplateTest < Minitest::Test
   def test_version_matches_release
-    assert_equal "0.4.7", ::RecordingStudioWordpressPluginTemplate::VERSION
+    assert_equal "0.4.8", ::RecordingStudioWordpressPluginTemplate::VERSION
   end
 
   def test_engine_exists
@@ -151,6 +151,7 @@ class RecordingStudioWordpressPluginTemplateTest < Minitest::Test
     assert_includes helper_source, "Home"
     assert_includes helper_source, "Recordings tree"
     assert_includes helper_source, "Registered apps"
+    assert_includes helper_source, "API Keys"
   end
 
   def test_dummy_login_layout_keeps_flatpack_assets_without_tight_main_offset
@@ -304,6 +305,7 @@ class RecordingStudioWordpressPluginTemplateTest < Minitest::Test
 
     assert_includes view_source, 'title: "WordPress widgets host"'
     assert_includes view_source, "Registered apps"
+    assert_includes view_source, "API Keys"
     refute_includes view_source, "Nothing talks to WordPress yet"
     assert_includes view_source, "FlatPack::Card::Component"
     assert_includes view_source, "dummy_page_nav"